+++ R/00_GridSearch.R
@@ -48,7 +48,17 @@ GridSearch <- function(
     fun = "Biclustering",
     index = "BIC",
     ...) {
+  obj <- dataFormat(obj)
+  testlength <- NCOL(obj$U)
+  nobs <- NROW(obj$U)
+  # ------------------------------------------ Biclustering
   if (fun == "Biclustering") {
+    if (max_ncls >= nobs) {
+      max_ncls <- nobs - 1
+    }
+    if (max_nfld >= testlength) {
+      max_nfld <- testlength - 1
+    }
     ret <- matrix(NA, nrow = length(2:max_ncls), ncol = length(2:max_nfld))
     colnames(ret) <- paste("nfld", 2:max_nfld)
     rownames(ret) <- paste("ncls", 2:max_ncls)
@@ -108,8 +118,10 @@ GridSearch <- function(
     } else if (index %in% maximize_indices) {
       optimal_idx <- which(ret == max(ret, na.rm = TRUE), arr.ind = TRUE)
     } else {
-      stop("Unknown index: ", index, ". Please specify one of: ",
-           paste(c(minimize_indices, maximize_indices), collapse = ", "))
+      stop(
+        "Unknown index: ", index, ". Please specify one of: ",
+        paste(c(minimize_indices, maximize_indices), collapse = ", ")
+      )
     }
     optimal_ncls <- optimal_idx[1] + 1
     optimal_nfld <- optimal_idx[2] + 1
@@ -149,7 +161,11 @@ GridSearch <- function(
       optimal_result = optimal_result,
       failed_settings = failed_settings
     )
+    # ------------------------------------------ LCA / LRA
   } else if (fun == "LCA" || fun == "LRA") {
+    if (max_ncls >= nobs) {
+      max_ncls <- nobs - 1
+    }
     ret <- vector(length = length(2:max_ncls))
     names(ret) <- paste0("ncls", 2:max_ncls)
     extra_args <- list(...)
@@ -202,8 +218,10 @@ GridSearch <- function(
     } else if (index %in% maximize_indices) {
       optimal_ncls <- which.max(ret) + 1
     } else {
-      stop("Unknown index: ", index, ". Please specify one of: ",
-           paste(c(minimize_indices, maximize_indices), collapse = ", "))
+      stop(
+        "Unknown index: ", index, ". Please specify one of: ",
+        paste(c(minimize_indices, maximize_indices), collapse = ", ")
+      )
     }
 
     # Display warning for failed convergence